% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fZooMSS_Xtras.R
\name{fZooMSS_Biomass}
\alias{fZooMSS_Biomass}
\title{Convert ZooMSS abundance matrices to biomass by multiplying by body weights}
\usage{
fZooMSS_Biomass(res, vmdl)
}
\arguments{
\item{res}{List of abundance matrices from ZooMSS output}

\item{vmdl}{ZooMSS model object containing weight vector (param$w)}
}
\value{
List of biomass matrices in grams wet weight
}
\description{
Converts abundance data to wet weight biomass by multiplying abundances
by the corresponding body weights for each size class.
}
\details{
Convert Abundance to Biomass

This function transforms abundance matrices to biomass by applying the
weight vector across size classes. Essential for analyses requiring biomass
units rather than abundance counts.
}
