% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoomss_plotting.R
\name{plotEnvironment}
\alias{plotEnvironment}
\title{Plot environmental forcing data}
\usage{
plotEnvironment(env_data)
}
\arguments{
\item{env_data}{Environmental data frame with time, sst, chl columns}
}
\value{
ggplot object (if patchwork available) or list of two ggplot objects
}
\description{
Creates plots of sea surface temperature and chlorophyll time series
for visualizing environmental forcing data used in ZooMSS model runs.
}
\details{
Plot Environmental Time Series

This function creates two separate plots with different y-axes scales:
\itemize{
\item SST plot (red line) with temperature in deg C
\item Chlorophyll plot (green line) with concentration in mg/m^3
}

The plots can be combined using the patchwork package if available, otherwise
separate plots are returned as a list. This helps users visualize the
environmental forcing that drives ZooMSS model dynamics.
}
\examples{
# Create sample data and plot
env_data <- data.frame(
  time = 1:100,
  dt = 0.01,
  sst = 15 + 3*sin(2*pi*(1:100)/50),
  chl = 0.5 + 0.2*cos(2*pi*(1:100)/50)
)
plots <- plotEnvironment(env_data)

}
